// src/api/logix-plugin/models/PedidoAceiteCompraResumoDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type PedidoAceiteCompraResumoDto = {
  /**
   * @type string
   */
  codEmpresa: string;
  /**
   * @type string
   */
  numeroPedido: string;
  /**
   * @type string
   */
  numeroVersao: string;
  /**
   * @type object
   */
  dataEmissao: object | null;
  /**
   * @type object
   */
  comprador: object | null;
  /**
   * @type object
   */
  fornecedor: object | null;
  /**
   * @type object
   */
  valorTotal: object | null;
};